test: separate wallet and node HTTP tests, cover legacy and segwit
